Refactored some components
